refactor(core): split claude/history.ts into converters, subagents, tool-result modules
history.ts had grown to 731 lines (2.4x the 300-line file limit). Pure
extraction, no behavior change: entry-to-ChatMessage converters move to
history-converters.ts, subagent JSONL collection helpers move to
history-subagents.ts, and the tool-result-block builder (shared by
converting, subagent collection, and two external callers) moves to
history-tool-result.ts. history.ts keeps discovery/loadHistory orchestration.
