+++ packages/core/src/plugins/builtin/claude/history.ts
@@ -3,470 +3,24 @@ import { access, constants, readdir } from 'node:fs/promises';
 import { createInterface } from 'node:readline';
 import { homedir } from 'node:os';
 import path from 'node:path';
-import { nanoid } from 'nanoid';
-import type { ChatMessage, MessageContent, DiffHunk, SkillFileEntry } from '@qlan-ro/mainframe-types';
-
-export function deriveModifiedFile(
-  tur: Record<string, unknown> | undefined,
-  originalFile: string | undefined,
-): string | undefined {
-  if (!tur) return undefined;
-  if (typeof tur.content === 'string' && (tur.type === 'create' || tur.type === 'update')) {
-    return tur.content;
-  }
-  if (originalFile && typeof tur.oldString === 'string') {
-    const oldStr = tur.oldString;
-    const newStr = (tur.newString as string) ?? '';
-    return tur.replaceAll ? originalFile.split(oldStr).join(newStr) : originalFile.replace(oldStr, newStr);
-  }
-  return undefined;
-}
-
+import type { ChatMessage, MessageContent, SkillFileEntry } from '@qlan-ro/mainframe-types';
 import { resolveSkillPath } from './skill-path.js';
 import { createChildLogger } from '../../../logger.js';
+import {
+  convertHistoryEntry,
+  synthesizeUnknownCommandFromUserEntry,
+  synthesizeSkillLoadedFromUserEntry,
+} from './history-converters.js';
+import {
+  collectAgentProgressTools,
+  collectSubagentToolResults,
+  captureAgentIdMapping,
+  collectSubagentAssistantBlocks,
+  attachSubagentToolResults,
+  injectAgentChildren,
+} from './history-subagents.js';
 
 const log = createChildLogger('claude:history');
-
-/**
- * isMeta user entries whose first text block starts with
- * "Base directory for this skill: <dir>" are the CLI's skill-content
- * injections. We keep them out of the chat transcript (as before) but
- * synthesize a transient system message with a `skill_loaded` block so
- * SkillLoadedCard renders on history replay.
- */
-/**
- * "Unknown command: /X" user entries are CLI feedback — the CLI never writes
- * the original user-typed /X to JSONL, so on replay the typed bubble is lost.
- * Synthesize both components: the invocation bubble and the error pill, so
- * history mirrors what the user saw live.
- */
-function synthesizeUnknownCommandFromUserEntry(entry: Record<string, unknown>, chatId: string): ChatMessage[] | null {
-  const message = entry.message as { content?: unknown } | undefined;
-  const content = message?.content;
-  if (typeof content !== 'string') return null;
-  // Match both CLI error variants. "Unknown command:" comes from the
-  // MalformedCommandError path (processSlashCommand.tsx:820) and retains the
-  // leading slash. "Unknown skill:" comes from the !hasCommand path
-  // (processSlashCommand.tsx:347) and omits the slash — we add it back so the
-  // synthesized user bubble consistently reads like "/foo".
-  const match = /^Unknown (?:command|skill):\s+\/?(\S+)/.exec(content.trim());
-  if (!match?.[1]) return null;
-  const cmd = `/${match[1]}`;
-  const uuid = (entry.uuid as string) ?? nanoid();
-  const timestamp = (entry.timestamp as string) ?? new Date().toISOString();
-  return [
-    {
-      id: `unknown-cmd-user-${uuid}`,
-      chatId,
-      type: 'user',
-      content: [{ type: 'text', text: cmd }],
-      timestamp,
-    },
-    {
-      id: `unknown-cmd-err-${uuid}`,
-      chatId,
-      type: 'system',
-      content: [{ type: 'text', text: content.trim() }],
-      timestamp,
-    },
-  ];
-}
-
-function synthesizeSkillLoadedFromUserEntry(entry: Record<string, unknown>, chatId: string): ChatMessage | null {
-  const message = entry.message as { content?: unknown } | undefined;
-  const content = message?.content;
-  if (!Array.isArray(content) || content.length === 0) return null;
-  const first = content[0] as { type?: string; text?: string } | undefined;
-  if (first?.type !== 'text' || typeof first.text !== 'string') return null;
-  const match = /^Base directory for this skill:\s*(.+?)(?:\n|$)/m.exec(first.text);
-  if (!match?.[1]) return null;
-  const baseDir = match[1].trim();
-  const skillName = path.basename(baseDir);
-  const skillPath = path.extname(baseDir) ? baseDir : path.join(baseDir, 'SKILL.md');
-  const skillContent = first.text.replace(/^Base directory for this skill:[^\n]*\n?/m, '').trim();
-  const uuid = (entry.uuid as string) ?? nanoid();
-  return {
-    id: `skill-loaded-${uuid}`,
-    chatId,
-    type: 'system',
-    content: [{ type: 'skill_loaded', skillName, path: skillPath, content: skillContent }],
-    timestamp: (entry.timestamp as string) ?? new Date().toISOString(),
-  };
-}
-
-export function extractToolResultContent(content: unknown): string {
-  if (typeof content === 'string') return content;
-  if (Array.isArray(content)) {
-    const texts: string[] = [];
-    for (const block of content) {
-      if (typeof block === 'object' && block !== null && 'text' in block && typeof block.text === 'string') {
-        texts.push(block.text);
-      }
-    }
-    if (texts.length > 0) return texts.join('\n');
-  }
-  return JSON.stringify(content ?? '');
-}
-
-/**
- * Extract text/image content blocks from a raw user-role block array (JSONL
- * `message.content` or a queued-command `prompt`). The two call sites differ
- * on empty text: history replay keeps empty text blocks and drops CLI
- * interrupt markers (`skipInterrupted`), while queued-command prompts drop
- * empty/whitespace-only text and never carry an interrupt marker.
- */
-function extractUserContentBlocks(
-  blocks: Array<Record<string, unknown>>,
-  opts: { skipInterrupted?: boolean } = {},
-): MessageContent[] {
-  const result: MessageContent[] = [];
-  for (const block of blocks) {
-    if (block.type === 'text') {
-      const text = block.text as string | undefined;
-      if (opts.skipInterrupted) {
-        const t = text || '';
-        if (!t.startsWith('[Request interrupted')) result.push({ type: 'text', text: t });
-      } else if (typeof text === 'string' && text.trim()) {
-        result.push({ type: 'text', text });
-      }
-    } else if (block.type === 'image') {
-      const source = block.source as Record<string, unknown> | undefined;
-      if (source?.type === 'base64') {
-        result.push({ type: 'image', mediaType: source.media_type as string, data: source.data as string });
-      }
-    }
-  }
-  return result;
-}
-
-export function buildToolResultBlocks(
-  message: Record<string, unknown>,
-  tur: Record<string, unknown> | undefined,
-): MessageContent[] {
-  const rawContent = message.content;
-  if (!Array.isArray(rawContent)) return [];
-
-  const sp = tur?.structuredPatch as DiffHunk[] | undefined;
-  const originalFile = tur?.originalFile as string | undefined;
-  const modifiedFile = deriveModifiedFile(tur, originalFile);
-
-  const blocks: MessageContent[] = [];
-  for (const block of rawContent) {
-    if (block.type !== 'tool_result') continue;
-    blocks.push({
-      type: 'tool_result',
-      toolUseId: (block.tool_use_id as string) || '',
-      content: extractToolResultContent(block.content),
-      isError: !!block.is_error,
-      ...(sp?.length ? { structuredPatch: sp } : {}),
-      ...(originalFile != null ? { originalFile } : {}),
-      ...(modifiedFile != null ? { modifiedFile } : {}),
-    });
-  }
-  return blocks;
-}
-
-function convertUserEntry(
-  entry: Record<string, unknown>,
-  message: Record<string, unknown>,
-  chatId: string,
-): ChatMessage | null {
-  const rawContent = message.content;
-  const contentBlocks: MessageContent[] = [];
-  const toolUseResult = entry.toolUseResult as Record<string, unknown> | undefined;
-
-  if (typeof rawContent === 'string') {
-    // String rawContent: user-typed text stored by Claude CLI when message.content
-    // is not an array. History must render it; live stream doesn't re-emit it
-    // (sendMessage() already created that ChatMessage).
-    // Known internal strings are filtered below — add new patterns here if they
-    // appear in JSONL but should never render in UI.
-    // TODO(task-support): render <task-notification> content once task UI is ready
-    if (rawContent.startsWith('<task-notification>')) return null;
-    contentBlocks.push({ type: 'text', text: rawContent });
-  } else if (Array.isArray(rawContent)) {
-    // Tool results — use shared builder (same logic as live stream)
-    const toolResults = buildToolResultBlocks(message, toolUseResult);
-    contentBlocks.push(...toolResults);
-
-    // Text and image blocks are intentionally only in history:
-    // live stream doesn't re-emit them because sendMessage() already created
-    // the user ChatMessage and tool results come via a separate tool_result entry.
-    contentBlocks.push(
-      ...extractUserContentBlocks(rawContent as Array<Record<string, unknown>>, { skipInterrupted: true }),
-    );
-  }
-
-  if (contentBlocks.length === 0) return null;
-
-  const hasToolResult = contentBlocks.some((b) => b.type === 'tool_result');
-  return {
-    id: (entry.uuid as string) || nanoid(),
-    chatId,
-    type: hasToolResult ? 'tool_result' : 'user',
-    content: contentBlocks,
-    timestamp: (entry.timestamp as string) || new Date().toISOString(),
-    metadata: { source: 'history' },
-  };
-}
-
-function convertAssistantEntry(
-  entry: Record<string, unknown>,
-  message: Record<string, unknown>,
-  chatId: string,
-): ChatMessage | null {
-  const rawContent = message.content;
-  const contentBlocks: MessageContent[] = [];
-
-  if (Array.isArray(rawContent)) {
-    for (const block of rawContent) {
-      if (block.type === 'text') {
-        contentBlocks.push({ type: 'text', text: block.text || '' });
-      } else if (block.type === 'thinking') {
-        contentBlocks.push({ type: 'thinking', thinking: block.thinking || '' });
-      } else if (block.type === 'tool_use') {
-        contentBlocks.push({
-          type: 'tool_use',
-          id: block.id || '',
-          name: block.name || '',
-          input: (block.input as Record<string, unknown>) || {},
-        });
-      }
-    }
-  }
-
-  if (contentBlocks.length === 0) return null;
-
-  const meta: Record<string, unknown> = { source: 'history' };
-  if (message.model) meta.model = message.model;
-  const usage = message.usage as
-    | {
-        input_tokens?: number;
-        output_tokens?: number;
-        cache_creation_input_tokens?: number;
-        cache_read_input_tokens?: number;
-      }
-    | undefined;
-  if (usage) meta.usage = usage;
-
-  return {
-    id: (entry.uuid as string) || nanoid(),
-    chatId,
-    type: 'assistant',
-    content: contentBlocks,
-    timestamp: (entry.timestamp as string) || new Date().toISOString(),
-    metadata: meta,
-  };
-}
-
-/**
- * A queued message the CLI drained mid-turn persists to JSONL as a structured
- * `attachment` entry (type queued_command), NOT as a user entry — verified
- * against CLI 2.1.198 on the stream-json + --replay-user-messages path
- * (docs/adapters/claude/QUEUE.md, local capture 2026-07-04). Converting it
- * here is what keeps mid-turn-drained messages visible after reload, at their
- * consumption point.
- */
-function convertQueuedCommandEntry(entry: Record<string, unknown>, chatId: string): ChatMessage | null {
-  const attachment = entry.attachment as Record<string, unknown> | undefined;
-  if (attachment?.type !== 'queued_command' || attachment.commandMode !== 'prompt') return null;
-
-  const prompt = attachment.prompt;
-  const contentBlocks: MessageContent[] = [];
-  if (typeof prompt === 'string') {
-    if (prompt.trim()) contentBlocks.push({ type: 'text', text: prompt });
-  } else if (Array.isArray(prompt)) {
-    contentBlocks.push(...extractUserContentBlocks(prompt as Array<Record<string, unknown>>));
-  }
-  if (contentBlocks.length === 0) return null;
-
-  return {
-    id: (entry.uuid as string) || nanoid(),
-    chatId,
-    type: 'user',
-    content: contentBlocks,
-    timestamp: (entry.timestamp as string) || (attachment.timestamp as string) || new Date().toISOString(),
-    metadata: { source: 'history' },
-  };
-}
-
-export function convertHistoryEntry(entry: Record<string, unknown>, chatId: string): ChatMessage | null {
-  const type = entry.type as string;
-
-  if (type === 'attachment') return convertQueuedCommandEntry(entry, chatId);
-
-  if (type === 'system' && entry.subtype === 'compact_boundary') {
-    return {
-      id: (entry.uuid as string) || nanoid(),
-      chatId,
-      type: 'system',
-      content: [{ type: 'compaction' }],
-      timestamp: (entry.timestamp as string) || new Date().toISOString(),
-      metadata: { source: 'history', internal: true },
-    };
-  }
-
-  if (type === 'result' && entry.subtype === 'error_during_execution' && entry.is_error !== false) {
-    return {
-      id: (entry.uuid as string) || nanoid(),
-      chatId,
-      type: 'error',
-      content: [{ type: 'error', message: 'Session ended unexpectedly' }],
-      timestamp: (entry.timestamp as string) || new Date().toISOString(),
-      metadata: { source: 'history' },
-    };
-  }
-
-  if (type !== 'user' && type !== 'assistant') return null;
-
-  const message = entry.message as Record<string, unknown> | undefined;
-  if (!message) return null;
-
-  if (type === 'user') return convertUserEntry(entry, message, chatId);
-  if (type === 'assistant') return convertAssistantEntry(entry, message, chatId);
-  return null;
-}
-
-/**
- * Flatten a subagent assistant message's content (tool_use / text / thinking)
- * onto the accumulated child-block list keyed by parentId. Shared by the two
- * collection paths (live `data.message` progress entries and subagent JSONL
- * assistant entries), which derive parentId/content differently but append the
- * blocks identically.
- */
-function appendAssistantBlocks(
-  parentId: string,
-  content: Array<Record<string, unknown>>,
-  agentTools: Map<string, MessageContent[]>,
-): void {
-  const existing = agentTools.get(parentId) ?? [];
-  for (const block of content) {
-    if (block.type === 'tool_use') {
-      existing.push({
-        type: 'tool_use',
-        id: (block.id as string) || nanoid(),
-        name: block.name as string,
-        input: (block.input as Record<string, unknown>) ?? {},
-        parentToolUseId: parentId,
-      });
-    } else if (block.type === 'text') {
-      const text = (block.text as string) || '';
-      if (text.trim()) existing.push({ type: 'text', text, parentToolUseId: parentId });
-    } else if (block.type === 'thinking') {
-      const t = (block.thinking as string) || '';
-      if (t.trim()) existing.push({ type: 'thinking', thinking: t, parentToolUseId: parentId });
-    }
-  }
-  if (existing.length > 0) agentTools.set(parentId, existing);
-}
-
-function collectAgentProgressTools(entry: Record<string, unknown>, agentTools: Map<string, MessageContent[]>): void {
-  const parentId = entry.parentToolUseID as string | undefined;
-  if (!parentId) return;
-  const data = entry.data as Record<string, unknown>;
-  const msg = data.message as Record<string, unknown> | undefined;
-  const inner = msg?.message as Record<string, unknown> | undefined;
-  if (!inner || inner.role !== 'assistant') return;
-  const content = inner.content as Array<Record<string, unknown>> | undefined;
-  if (!Array.isArray(content)) return;
-
-  appendAssistantBlocks(parentId, content, agentTools);
-}
-
-/** Extract tool_result blocks from subagent JSONL user entries. */
-function collectSubagentToolResults(
-  entry: Record<string, unknown>,
-  results: Map<string, MessageContent & { type: 'tool_result' }>,
-): void {
-  if (entry.type !== 'user') return;
-  const message = entry.message as Record<string, unknown> | undefined;
-  if (!message) return;
-  const rawContent = message.content;
-  if (!Array.isArray(rawContent)) return;
-  const toolUseResult = entry.toolUseResult as Record<string, unknown> | undefined;
-  const blocks = buildToolResultBlocks(message, toolUseResult);
-  for (const block of blocks) {
-    if (block.type === 'tool_result') {
-      results.set(block.toolUseId, block);
-    }
-  }
-}
-
-/**
- * Capture the agentId → parent tool_use_id mapping from a parent-JSONL user
- * entry whose tool_result corresponds to a Task/Agent dispatch. CLI 2.1.118+
- * does not write parentToolUseID into subagent JSONL entries; this map is the
- * only way to link a subagent's blocks back to its dispatching tool_use.
- */
-function captureAgentIdMapping(entry: Record<string, unknown>, map: Map<string, string>): void {
-  if (entry.type !== 'user') return;
-  const tur = (entry.toolUseResult ?? entry.tool_use_result) as Record<string, unknown> | undefined;
-  const agentId = typeof tur?.agentId === 'string' ? (tur.agentId as string) : undefined;
-  if (!agentId) return;
-  const message = entry.message as { content?: unknown } | undefined;
-  const content = message?.content;
-  if (!Array.isArray(content)) return;
-  for (const block of content as Array<Record<string, unknown>>) {
-    if (block.type === 'tool_result' && typeof block.tool_use_id === 'string') {
-      map.set(agentId, block.tool_use_id);
-      return;
-    }
-  }
-}
-
-/** Collect assistant text/thinking/tool_use blocks from subagent JSONL assistant entries. */
-function collectSubagentAssistantBlocks(
-  entry: Record<string, unknown>,
-  agentTools: Map<string, MessageContent[]>,
-  agentIdMap?: Map<string, string>,
-): void {
-  let parentId = entry.parentToolUseID as string | undefined;
-  if (!parentId && agentIdMap) {
-    const agentId = entry.agentId as string | undefined;
-    if (agentId) parentId = agentIdMap.get(agentId);
-  }
-  if (!parentId) return;
-  if (entry.type !== 'assistant') return;
-  const message = entry.message as Record<string, unknown> | undefined;
-  const content = message?.content as Array<Record<string, unknown>> | undefined;
-  if (!Array.isArray(content)) return;
-
-  appendAssistantBlocks(parentId, content, agentTools);
-}
-
-/** Inject subagent tool_result blocks after their matching tool_use in assistant messages. */
-function attachSubagentToolResults(
-  messages: ChatMessage[],
-  results: Map<string, MessageContent & { type: 'tool_result' }>,
-): void {
-  for (const msg of messages) {
-    if (msg.type !== 'assistant') continue;
-    const newContent: MessageContent[] = [];
-    for (const block of msg.content) {
-      newContent.push(block);
-      if (block.type === 'tool_use') {
-        const toolResult = results.get(block.id);
-        if (toolResult) newContent.push({ ...toolResult, parentToolUseId: block.parentToolUseId });
-      }
-    }
-    msg.content = newContent;
-  }
-}
-
-function injectAgentChildren(messages: ChatMessage[], agentTools: Map<string, MessageContent[]>): void {
-  for (const msg of messages) {
-    if (msg.type !== 'assistant') continue;
-    const newContent: MessageContent[] = [];
-    for (const block of msg.content) {
-      newContent.push(block);
-      if (block.type === 'tool_use' && (block.name === 'Agent' || block.name === 'Task')) {
-        const children = agentTools.get(block.id);
-        if (children) newContent.push(...children);
-      }
-    }
-    msg.content = newContent;
-  }
-}
 
 function getSessionJsonlPath(sessionId: string, projectPath: string): { jsonlPath: string; projectDir: string } {
   const encodedPath = projectPath.replace(/[^a-zA-Z0-9-]/g, '-');
